separate container and presentation
