+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/Categories';
+import App from './components/Dashboard';
 import './main.css';
 
 ReactDOM.render(
